Separate start up files
