+++ src/applications/vaos/vaos-entry.jsx
@@ -1,7 +1,7 @@
 import 'platform/polyfills';
 import './sass/vaos.scss';
 
-import { startReactRouterDOMApp as startApp } from 'platform/startup';
+import startApp from 'platform/startup/router';
 import createRoutesWithStore from './routes';
 import manifest from './manifest.json';
 import reducer from './reducers';
